+++ Package.swift
@@ -25,13 +25,13 @@ let package = Package(
     targets: [
         .binaryTarget(
             name: "SendbirdUIKit",
-            url: "https://github.com/sendbird/sendbird-uikit-ios/releases/download/3.33.0/SendbirdUIKit.xcframework.zip", // SendbirdUIKit_URL
-            checksum: "faefdaa29859b20c9268c15f77482e121976196012c0879ffe3572b27a5e83ce" // SendbirdUIKit_CHECKSUM
+            url: "https://github.com/sendbird/sendbird-uikit-ios/releases/download/3.33.1/SendbirdUIKit.xcframework.zip", // SendbirdUIKit_URL
+            checksum: "b55c89d32f9634270be305c92fc67c0c0b04c23328d4f9436e55037dce3ec9ca" // SendbirdUIKit_CHECKSUM
         ),
         .binaryTarget(
             name: "SendbirdUIMessageTemplate",
-            url: "https://github.com/sendbird/sendbird-uikit-ios/releases/download/3.33.0/SendbirdUIMessageTemplate.xcframework.zip", // SendbirdUIMessageTemplate_URL
-            checksum: "3c429a0416ebfe9f09b529844371774660939a70139abe6d3781c03eb46724d5" // SendbirdUIMessageTemplate_CHECKSUM
+            url: "https://github.com/sendbird/sendbird-uikit-ios/releases/download/3.33.1/SendbirdUIMessageTemplate.xcframework.zip", // SendbirdUIMessageTemplate_URL
+            checksum: "99be28b7f9e23a5749f28565362a335e709d2752fada6af15032879882f69542" // SendbirdUIMessageTemplate_CHECKSUM
         ),
         .target(
             name: "SendbirdUIKitTarget",